+++ app/src/main/java/com/droppyn/domain/Size.kt
@@ -2,7 +2,7 @@ package com.droppyn.domain
 
 data class Size (
 	val id : String,
-	val us : Int,
+	val us : Double,
 	val uk : Double,
 	val eu : String,
 	val type : String,
