+++ app/src/main/java/com/droppyn/network/dto/NetworkSizeContainer.kt
@@ -2,7 +2,7 @@ package com.droppyn.network.dto
 
 data class SizeDTO (
 	val id : String,
-	val us : Int,
+	val us : Double,
 	val uk : Double,
 	val eu : String,
 	val type : String,
